pyproject.toml
[tool.poetry]
name = "starke"
version = "0.1.0"
description = "Relatório Diário de Fluxo de Caixa - Sistema automatizado de coleta, processamento e envio de relatórios"
authors = ["Starke <tech@starke.com>"]
readme = "README.md"
packages = [{include = "starke", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Core
pydantic = "^2.0"
pydantic-settings = "^2.0"
# HTTP & API
httpx = "^0.27"
tenacity = "^8.2"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.32"}
python-jose = {extras = ["cryptography"], version = "^3.3"}
passlib = {extras = ["bcrypt"], version = "^1.7"}
bcrypt = ">=4.0.0,<5.0.0"
python-multipart = "^0.0.9"
itsdangerous = "^2.1"
# Database
sqlalchemy = "^2.0"
alembic = "^1.13"
psycopg2-binary = "^2.9"
# Templates & Reports
jinja2 = "^3.1"
# Email
aiosmtplib = "^3.0"
# Utilities
python-dateutil = "^2.8"
pytz = "^2024.1"
structlog = "^24.0"
click = "^8.1"
pyyaml = "^6.0"
# Scheduler
apscheduler = "^3.10"
# Email validation
email-validator = "^2.0"
# AWS S3
boto3 = "^1.35"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.1"
pytest-asyncio = "^0.23"
pytest-mock = "^3.12"
ruff = "^0.3"
mypy = "^1.9"
black = "^24.0"
ipython = "^8.0"

[tool.poetry.scripts]
starke = "starke.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["starke"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/starke --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
